feat: Make 'pup deploy' idempotent (#101)
Make 'pup deploy', and any sub-command run singularly, idempotent.
Several user prompts have been added to allow using existing
data/environments or to re-initialize.

Summary of most significant changes:

- Hardware validation prompts user to skip if an inventory exists with
  MAC addresses already populated.

- Python virtual environments on deployer and inside container are
  created using 'scripts/venv_install.sh'. Both are named 'pup-env'.

- Container files are copied through 'lxc-attach' instead of SSH. When
  re-running with an existing container only newer files will be
  transferred.

- BMCs are only issued a cold reset if the IP address changes.

+++ scripts/python/lib/container.py
@@ -33,8 +33,8 @@ import lxc
 import lib.logger as logger
 from lib.config import Config
 from lib.exception import UserException
-from lib.ssh import SSH_CONNECTION, SSH_Exception
 import lib.genesis as gen
+from lib.utilities import bash_cmd
 
 
 class Container(object):
@@ -70,7 +70,6 @@ class Container(object):
         self.cont_playbooks_path = gen.get_container_playbooks_path()
         self.depl_package_path = gen.get_package_path()
         self.depl_python_path = gen.get_python_path()
-        self.depl_os_images_path = gen.get_os_images_path()
         self.depl_playbooks_path = gen.get_playbooks_path()
         self.config_file = gen.get_config_file_name()
 
@@ -95,73 +94,6 @@ class Container(object):
         # Get a file descriptor for stdout
         self.fd = open(os.path.join(gen.GEN_LOGS_PATH,
                                     self.name + '.stdout.log'), 'w')
-
-    def open_ssh(self):
-        cont_ipaddr = self.cont.get_ips(
-            interface='eth0', family='inet', timeout=5)[0]
-        try:
-            ssh = SSH_CONNECTION(
-                cont_ipaddr,
-                username='root',
-                key_filename=self.PRIVATE_SSH_KEY_FILE)
-        except SSH_Exception as exc:
-            msg = "SSH to container '{}' at '{}' failed - {}".format(
-                self.name, cont_ipaddr, exc)
-            self.log.error(msg)
-            raise UserException(msg)
-        return ssh
-
-    def open_sftp(self, ssh):
-        try:
-            return ssh.open_sftp_session()
-        except Exception as exc:
-            error = "Failed to open sftp session to the '{}' container - {}"
-            error = error.format(self.name, exc)
-            self.log.error(error)
-            raise UserException(error)
-        self.log.debug("Opened sftp session to the '{}' container".format(
-            self.name))
-
-    def _close_ssh(self, ssh):
-        try:
-            ssh.close()
-        except Exception as exc:
-            error = "Failed to close sftp session to the '{}' container - {}"
-            error = error.format(self.name, exc)
-            self.log.error(error)
-            raise UserException(error)
-        self.log.debug("Closed sftp session to the '{}' container".format(
-            self.name))
-
-    def _mkdir_sftp(self, sftp, dir_):
-        try:
-            sftp.chdir(dir_)
-        except IOError:
-            try:
-                sftp.mkdir(dir_)
-            except Exception as exc:
-                error = (
-                    "Failed via sftp to create the '{}' directory in the '{}'"
-                    " container - {}")
-                error = error.format(dir_, self.name, exc)
-                self.log.error(error)
-                raise UserException(error)
-            msg = "Created via sftp the '{}' directory in the '{}' container"
-            self.log.debug(msg.format(dir_, self.name))
-
-    def _copy_sftp(self, sftp, src, dst):
-        try:
-            sftp.put(src, dst)
-        except Exception as exc:
-            error = (
-                "Failed via sftp to copy '{}' to '{}' in the '{}' container"
-                " - {}")
-            error = error.format(src, dst, self.name, exc)
-            self.log.error(error)
-            raise UserException(error)
-        self.log.debug(
-            "Copied via sftp '{}' to '{}' in the '{}' container".format(
-                src, dst, self.name))
 
     def check_permissions(self, user):
         # Enumerate LXC bridge
@@ -233,7 +165,6 @@ class Container(object):
         if rc:
             error = "Failed running '{}' in the container '{}'".format(
                 ' '.join(cmd), self.name)
-            self.log.error(error)
             raise UserException(error)
         self.log.debug(
             "Successfully ran '{}' in the container '{}'".format(
@@ -243,22 +174,28 @@ class Container(object):
         # Check if container already exists
         if self.cont.defined:
             msg = "Container '%s' already exists" % self.name
-            self.log.error(msg)
-            raise UserException(msg)
-
-        # Create container
-        if not self.cont.create('download', lxc.LXC_CREATE_QUIET, self.rootfs):
-            msg = "Failed to create container '%s'" % self.name
-            self.log.error(msg)
-            raise UserException(msg)
-        self.log.debug("Created container '%s'" % self.name)
+            self.log.warning(msg)
+            print("\nPress enter to continue with node configuration using ")
+            print("existing container, or 'T' to terminate.")
+            resp = raw_input("\nEnter or 'T': ")
+            if resp == 'T':
+                sys.exit('POWER-Up stopped at user request')
+        else:
+            # Create container
+            if not self.cont.create('download', lxc.LXC_CREATE_QUIET,
+                                    self.rootfs):
+                msg = "Failed to create container '%s'" % self.name
+                self.log.error(msg)
+                raise UserException(msg)
+            self.log.debug("Created container '%s'" % self.name)
 
         # Start container
-        if not self.cont.start():
-            msg = "Failed to start container '%s'" % self.name
-            self.log.error(msg)
-            raise UserException(msg)
-        self.log.debug("Started container '%s'" % self.name)
+        if not self.cont.running:
+            if not self.cont.start():
+                msg = "Failed to start container '%s'" % self.name
+                self.log.error(msg)
+                raise UserException(msg)
+            self.log.debug("Started container '%s'" % self.name)
 
         # Get nameservers from /etc/resolv.conf outside container
         nameservers = []
@@ -277,9 +214,12 @@ class Container(object):
         # Update '/etc/resolv.conf' in container by updating
         # '/etc/resolvconf/resolv.conf.d/base'
         for line in nameservers:
-            entry = 'a|%s' % line
+            entry = '"a|%s"' % line
+            line = '"%s"' % line
             self.run_command(
-                ['ex', '-sc', entry, '-cx', self.RESOLV_CONF_BASE],
+                ['sh', '-c',
+                 'grep ' + line + ' ' + self.RESOLV_CONF_BASE + ' || '
+                 'ex  -sc ' + entry + ' -cx ' + self.RESOLV_CONF_BASE],
                 stdout=self.fd)
 
         # Sleep to allow /etc/resolv.conf to update
@@ -288,14 +228,17 @@ class Container(object):
 
         # Create user
         self.run_command(
-            ['adduser', '--disabled-password', '--gecos', 'GECOS', 'deployer'],
+            ['sh', '-c',
+             'grep deployer /etc/passwd || '
+             'adduser --disabled-password --gecos GECOS deployer'],
             stdout=self.fd)
 
         # Create '/root/.ssh' directory
-        self.run_command(['mkdir', '/root/.ssh'], stdout=self.fd)
+        self.run_command(['mkdir', '-p', '/root/.ssh'], stdout=self.fd)
 
         # Create '/root/.ssh/authorized_keys' file
-        self.run_command(['touch', '/root/.ssh/authorized_keys'], stdout=self.fd)
+        self.run_command(['touch', '/root/.ssh/authorized_keys'],
+                         stdout=self.fd)
 
         # Change '/root/.ssh' permissions to 0700
         self.run_command(['chmod', '700', '/root/.ssh'], stdout=self.fd)
@@ -304,21 +247,39 @@ class Container(object):
         self.run_command(['chmod', '600', '/root/.ssh/authorized_keys'],
                          stdout=self.fd)
 
-        key = RSA.generate(self.RSA_BIT_LENGTH)
-        # Create private ssh key
-        with open(self.PRIVATE_SSH_KEY_FILE, 'w') as ssh_key:
-            ssh_key.write(key.exportKey())
-        os.chmod(self.PRIVATE_SSH_KEY_FILE, 0o600)
-        # Create public ssh key
-        public_key = key.publickey().exportKey(format='OpenSSH')
-        with open(self.PUBLIC_SSH_KEY_FILE, 'w') as ssh_key:
-            ssh_key.write(public_key)
+        # Create new SSH private/public keys only if they don't exist
+        if (not os.path.isfile(self.PRIVATE_SSH_KEY_FILE) and
+                not os.path.isfile(self.PUBLIC_SSH_KEY_FILE)):
+            key = RSA.generate(self.RSA_BIT_LENGTH)
+            # Create private ssh key
+            with open(self.PRIVATE_SSH_KEY_FILE, 'w') as ssh_key:
+                ssh_key.write(key.exportKey())
+            os.chmod(self.PRIVATE_SSH_KEY_FILE, 0o600)
+            # Create public ssh key
+            public_key = key.publickey().exportKey(format='OpenSSH')
+            with open(self.PUBLIC_SSH_KEY_FILE, 'w') as ssh_key:
+                ssh_key.write(public_key)
+        # Throw exception if one of the key pair is missing
+        elif (not os.path.isfile(self.PRIVATE_SSH_KEY_FILE) and
+                os.path.isfile(self.PUBLIC_SSH_KEY_FILE)):
+            raise UserException("Private SSH key is missing but public exists")
+        elif (os.path.isfile(self.PRIVATE_SSH_KEY_FILE) and
+                not os.path.isfile(self.PUBLIC_SSH_KEY_FILE)):
+            raise UserException("Public SSH key is missing but private exists")
 
         # Add public ssh key to container
-        self.run_command([
-            'ex',
-            '-sc', 'a|%s' % public_key,
-            '-cx', '/root/.ssh/authorized_keys'], stdout=self.fd)
+        with open(self.PUBLIC_SSH_KEY_FILE, 'r') as file_in:
+            for line in file_in:
+                # public key file should only be 1 line
+                # if it's more than 1 the last will be used
+                public_key = line
+        entry = '"a|%s"' % public_key
+        line = '"%s"' % public_key
+        self.run_command(
+            ['sh', '-c',
+             'grep ' + line + ' /root/.ssh/authorized_keys || '
+             'ex  -sc ' + entry + ' -cx /root/.ssh/authorized_keys'],
+            stdout=self.fd)
 
         print()
         self.log.info('Installing software packages in container\n'
@@ -361,119 +322,38 @@ class Container(object):
                 cmd.append(pkg)
             self.run_command(cmd, stdout=self.fd)
 
-        # Install pip container packages
-        if ini.has_section(self.Packages.PIP.value):
-            cmd = ['pip', 'install']
-            for pkg in ini.options(self.Packages.PIP.value):
-                cmd.append(pkg)
-                print('.', end="")
-                sys.stdout.flush()
-            self.run_command(cmd, stdout=self.fd)
+        # Create project directory
+        self.run_command(['mkdir', '-p', self.cont_package_path],
+                         stdout=self.fd)
 
-        # Create project
-        self.run_command(['mkdir', self.cont_package_path], stdout=self.fd)
-
-        # Create virtual environment
-        self.run_command([
-            'virtualenv',
-            '--no-wheel',
-            '--system-site-packages',
-            self.cont_venv_path], stdout=self.fd)
-        # Open SSH to container
-        ssh = self.open_ssh()
-
-        # Install pip venv container packages
-        if ini.has_section(self.Packages.VENV.value):
-            for pkg, ver in ini.items(self.Packages.VENV.value):
-                cmd = [
-                    'source', self.cont_venv_path + '/bin/activate',
-                    '&&', 'pip', 'install']
-                cmd.append('{}=={}'.format(pkg, ver))
-                cmd.extend(['&&', 'deactivate'])
-                status, stdout_, stderr_ = ssh.send_cmd(' '.join(cmd))
-                if status:
-                    error = 'Failed venv pip install'
-                    self.log.error(error)
-                    self.log.error(' '.join(cmd))
-                    self.log.error(stderr_)
-                    raise UserException(error)
-                print('.', end="")
-                sys.stdout.flush()
-
-        # Open sftp session to container
-        sftp = self.open_sftp(ssh)
-
-        # Copy private ssh key to container
-        self._mkdir_sftp(sftp, self.cont_scripts_path)
-        self._copy_sftp(
-            sftp,
-            self.PRIVATE_SSH_KEY_FILE,
-            '/root/.ssh/gen')
-        print('.', end="")
-        sys.stdout.flush()
+        # Copy private ssh key pair to container
+        bash_cmd("cat {}  | lxc-attach -n {} -- /bin/bash -c \"cat > "
+                 "/root/.ssh/gen\"".format(self.PRIVATE_SSH_KEY_FILE,
+                                           self.name))
+        bash_cmd("cat {}  | lxc-attach -n {} -- /bin/bash -c \"cat > "
+                 "/root/.ssh/gen.pub\"".format(self.PUBLIC_SSH_KEY_FILE,
+                                               self.name))
 
         # Change private key file permissions to 0600
         self.run_command(['chmod', '600', '/root/.ssh/gen'], stdout=self.fd)
 
-        # Copy config file to container
-        self._copy_sftp(
-            sftp,
-            os.path.join(self.depl_package_path, self.config_file),
-            os.path.join(self.cont_package_path, self.config_file))
-        print('.', end="")
-        sys.stdout.flush()
+        # Copy power-up directory into container
+        bash_cmd("tar --exclude='inventory*.yml' --exclude='pup-venv' -h "
+                 "--exclude='logs' -C {} -c . | lxc-attach -n {} -- tar -C {} "
+                 "-xvp --keep-newer-files".format(self.depl_package_path,
+                                                  self.name,
+                                                  self.cont_package_path))
 
-        # Copy scripts/python directory to container
-        self._mkdir_sftp(sftp, self.cont_scripts_path)
-        self._mkdir_sftp(sftp, self.cont_python_path)
-        self.copy_dir_to_container(sftp, self.depl_python_path)
-        print('.', end="")
-        sys.stdout.flush()
-
-        # Add execute permission to dynamic inventory module
-        self.run_command([
-            'chmod',
-            'a+x',
-            os.path.join(self.cont_python_path, 'inventory.py')],
-            stdout=self.fd)
-
-        # Copy os_images directory to container
-        self._mkdir_sftp(sftp, self.cont_os_images_path)
-        self.copy_dir_to_container(sftp, self.depl_os_images_path)
-        print('.', end="")
-        sys.stdout.flush()
-
-        # Copy playbooks directory to container
-        self._mkdir_sftp(sftp, self.cont_playbooks_path)
-        self.copy_dir_to_container(sftp, self.depl_playbooks_path)
-        print('.', end="")
-        sys.stdout.flush()
+        # Install python virtual environment
+        self.run_command([self.cont_package_path + '/scripts/venv_install.sh',
+                          self.cont_package_path + '/'], stdout=self.fd)
 
         # Create file to indicate whether project is installed in a container
         self.run_command(['touch', self.cont_id_file], stdout=self.fd)
         print()
 
-        # Close ssh session to container
-        self._close_ssh(ssh)
-        self.fd.close()
-
-    def copy_dir_to_container(self, sftp, dir_local_path, follow_sym_link=True):
-        for dirpath, dirnames, filenames in os.walk(dir_local_path):
-            for dirname in dirnames:
-                self._mkdir_sftp(
-                    sftp,
-                    os.path.join(
-                        self.cont_package_path,
-                        os.path.relpath(dirpath, self.depl_package_path),
-                        dirname))
-            for filename in filenames:
-                filepath = os.path.join(dirpath, filename)
-                if os.path.islink(filepath) and not follow_sym_link:
-                    continue
-                self._copy_sftp(
-                    sftp,
-                    os.path.join(dirpath, filename),
-                    os.path.join(
-                        self.cont_package_path,
-                        os.path.relpath(dirpath, self.depl_package_path),
-                        filename))
+    def copy_dir_to_container(self, source_path, cont_dest_path):
+        bash_cmd("tar -h -C {} -c . | lxc-attach -n {} -- tar -C {} -xvp "
+                 "--keep-newer-files".format(source_path,
+                                             self.name,
+                                             cont_dest_path))
